requirements.txt


plotly==5.18.0

seaborn==0.13.0
matplotlib==3.8.2
Pillow==10.1.0
# --- App Framework ---
streamlit

# --- Google Cloud / BigQuery ---
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2

# --- Data Science / ML ---
pandas
numpy
scikit-learn
matplotlib
seaborn

# --- Optional utils ---
protobuf==4.25.3    # règle des problèmes BigQuery/Google API
pyarrow              # recommandé pour BigQuery et pandas